embedding blocked users page in settings and making blocked tags combobox always visible
