--- front-end/src/BlockedUsers.css
@@ -1,12 +1,0 @@
-.backToSettings{
-    transform: translatey(-75%);
-}
-.blockedUsersTitle{
-    text-align: center;
-    transform: translatey(-75%);
-}
-.blockedUsersPTag{
-    font-weight: 900;
-    font-size: 30px;
-    font-family: sans-serif;
-}
